Fix type bug and subtitle name bug when over 255

+++ internal/database/struct.go
@@ -410,7 +410,7 @@ type YouTubeVideoInfoStruct struct {
 		FormatID           string      `json:"format_id"`
 		FormatNote         string      `json:"format_note"`
 		SourcePreference   int         `json:"source_preference"`
-		Fps                int         `json:"fps"`
+		Fps                float64     `json:"fps"`
 		AudioChannels      interface{} `json:"audio_channels"`
 		Height             int         `json:"height"`
 		Quality            float64     `json:"quality"`
@@ -455,7 +455,7 @@ type YouTubeVideoInfoStruct struct {
 	Width          int         `json:"width"`
 	Height         int         `json:"height"`
 	Resolution     string      `json:"resolution"`
-	Fps            int         `json:"fps"`
+	Fps            float64     `json:"fps"`
 	DynamicRange   string      `json:"dynamic_range"`
 	Vcodec         string      `json:"vcodec"`
 	Vbr            float64     `json:"vbr"`
